# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QuoteCraft API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all saved quote templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Save a new quote template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a single quote template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteTemplate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a quote template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteTemplateInput"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteTemplate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a quote template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    QuoteLineItem:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: ["string", "null"]
        unit:
          type: string
        unitPrice:
          type: number
        quantity:
          type: number
        voiceKey:
          type: string
          description: Voice command keyword to set this field quantity
        overtimePercent:
          type: number
          description: >-
            Overtime markup percentage applied on top of the base unitPrice.
            Effective rate = unitPrice + (unitPrice * overtimePercent / 100).
            0 or absent means no overtime.
      required:
        - id
        - label
        - unit
        - unitPrice
        - quantity
        - voiceKey
    QuoteSettings:
      type: object
      properties:
        includeGst:
          type: boolean
        gstRate:
          type: number
        callOutFee:
          type: number
        publicHolidaySurchargePercent:
          type: number
        isPublicHoliday:
          type: boolean
        hasCallOut:
          type: boolean
      required:
        - includeGst
        - gstRate
        - callOutFee
        - publicHolidaySurchargePercent
        - isPublicHoliday
        - hasCallOut
    QuoteTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessDescription:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        settings:
          $ref: "#/components/schemas/QuoteSettings"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - businessDescription
        - lineItems
        - settings
        - createdAt
        - updatedAt
    QuoteTemplateInput:
      type: object
      properties:
        name:
          type: string
        businessDescription:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        settings:
          $ref: "#/components/schemas/QuoteSettings"
      required:
        - name
        - businessDescription
        - lineItems
        - settings
...